rustc: dismantle hir::util, mostly moving functions to methods.
